Additional Base Tests - yiisoft/data/pull/234 - msi 100
Applied  yiisoft/data/pull/234
Moved tests folder into psalm.xml - Level 1 tested
EntityReader's readOne function's limit(1) has been removed to get the msi to 100.

A suitable test needs to be developed to check for the IncrementInteger mutant on the readOne function with the mutant that changes limit(1) to limit(2) still causing a failure with current tests not picking it up mainly due to the \Generator->current returning mixed type where null, array, and object have to be checked in readOne with false also being returned.

+++ tests/Feature/Base/Writer/BaseEntityWriterTestCase.php
@@ -20,29 +20,34 @@ abstract class BaseEntityWriterTestCase extends TestCase
     public function testWrite(): void
     {
         $orm = $this->getOrm();
+        $entityWriter = $this->createEntityManager();
+        if (null !== $entityWriter) {
+            $writer = new EntityWriter($entityWriter);
+            $writer->write($users = [
+                $orm->make('user', ['number' => 99998, 'email' => 'super@test1.com', 'balance' => 1000.0]),
+                $orm->make('user', ['number' => 99999, 'email' => 'super@test2.com', 'balance' => 999.0]),
+            ]);
 
-        $writer = new EntityWriter($this->createEntityManager());
-        $writer->write($users = [
-            $orm->make('user', ['number' => 99998, 'email' => 'super@test1.com', 'balance' => 1000.0]),
-            $orm->make('user', ['number' => 99999, 'email' => 'super@test2.com', 'balance' => 999.0]),
-        ]);
-
-        $reader = new EntityReader(
-            $this->select('user')->where('number', 'in', [99998, 99999]),
-        );
-        $this->assertEquals($users, $reader->read());
+            $reader = new EntityReader(
+                $this->select('user')->where('number', 'in', [99998, 99999]),
+            );
+            $this->assertEquals($users, $reader->read());
+        }
     }
 
     public function testDelete(): void
     {
-        $writer = new EntityWriter($this->createEntityManager());
-        $reader = new EntityReader($this->select('user')->where('number', 'in', [1, 2, 3]));
-        // Iterator doesn't use cache
-        $entities = \iterator_to_array($reader->getIterator());
+        $entityWriter = $this->createEntityManager();
+        if (null !== $entityWriter) {
+            $writer = new EntityWriter($entityWriter);
+            $reader = new EntityReader($this->select('user')->where('number', 'in', [1, 2, 3]));
+            // Iterator doesn't use cache
+            $entities = \iterator_to_array($reader->getIterator());
 
-        $writer->delete($entities);
+            $writer->delete($entities);
 
-        $this->assertCount(3, $entities);
-        $this->assertEquals([], \iterator_to_array($reader->getIterator()));
+            $this->assertCount(3, $entities);
+            $this->assertEquals([], \iterator_to_array($reader->getIterator()));
+        }
     }
 }
